host as parameterized

+++ .cc/scripts/swagger.sh
@@ -1,6 +1,13 @@
 #!/bin/bash
 
-flaskswagger openbbs_middleware.main:app --host 173.255.216.176:3457 --base-path / --out-dir swagger --from-file-keyword=swagger_from_file --template ./apidoc/template.json
+if [ "$1" == "" ]; then
+    echo "usage: swagger.sh [host]"
+    exit 255
+fi
+
+host=$1
+
+flaskswagger openbbs_middleware.main:app --host ${host} --base-path / --out-dir swagger --from-file-keyword=swagger_from_file --template ./apidoc/template.json
 
 docker container stop swagger-ui
 docker container rm swagger-ui
